+++ frontend/public/sw.js
@@ -908,4 +908,4 @@ self.addEventListener("fetch", (event) => {
         })(),
     );
 });
-// Soundspan app build: "V6tKT0ZmMMNn7pr1fCsDc"
+// Soundspan app build: "jnhfucjBbsU86It4z0Q6i"